+++ include/f_pc/f_pc_leaf.h
@@ -30,7 +30,7 @@ typedef struct leaf_process_profile_definition {
     /* 0x3C */ leafdraw_method_class* mMethods;
 } leaf_process_profile_definition;
 
-s32 fpcLf_GetPriority(const leafdraw_class* pLeaf);
+s16 fpcLf_GetPriority(const leafdraw_class* pLeaf);
 s32 fpcLf_DrawMethod(leafdraw_method_class* pMthd, void* pUserData);
 s32 fpcLf_Draw(leafdraw_class* pMthd);
 s32 fpcLf_Execute(leafdraw_class* pLeaf);
